requirements.txt
requests
beautifulsoup4
pandas
selenium
python-dotenv